Add link to current NixOS Wiki and fix awkward wording (#149)
Currently, "1.2. Rationale for this series" starts out with the following sentence, which seems a bit odd:

The Nix, Nixpkgs, and NixOS manuals; and wiki are excellent resources for explaining how Nix/NixOS works, how you can use it, and the number of cool things being done with it.

Turns out 4d0c1d2 removes an item followed by a semicolon. This commit fixes the issue by removing the semicolon and also adds a link to the current NixOS Wiki.