build: Drop DocBook schema validation
It was not actually used because we did not pass it the manual.

And if we fix it by replacing the Bash `$combined` variable by a Nix one,
it will fail because we only have RelaxNG schema for Docbook 5.0,
which unlike later versions is too limited (e.g. does not allow
emphasis element within filename element).

- https://tdg.docbook.org/tdg/5.0/filename.html
- https://tdg.docbook.org/tdg/5.2/filename.html