07-working-derivation.md: use `derivation` instead of `mkDerivation`
When switching from the nix repl to a nix file, `derivation` was replaced with `pkgs.stdenv.mkDerivation`. This was not explained, and the examples work fine without this. It seems to me that when learning that you can define the derivation in a nix file instead of in the repl, there's no need to add another change.

I see that the next chapter uses plain `derivation` in the first examples, and then defines a new function `mkDerivation`. So I'm pretty certain that we should just use plain `derivation` here.